small cleanup
git-svn-id: https://svn.apache.org/repos/asf/ant/ivy/core/trunk@613802 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/ivy/plugins/repository/ssh/SshCache.java
@@ -327,10 +327,8 @@ public final class SshCache {
                 Message.verbose(":: SSH :: connected to " + host + "!");
                 setSession(username, host, port, session);
             } catch (JSchException e) {
-                if (passFile!=null) {
-                  if (passFile.exists()) {
+                if (passFile != null && passFile.exists()) {
                     passFile.delete();
-                  }
                 }
                 IOException ex = new IOException(e.getMessage());
                 ex.initCause(e);
